rename error to exception
