+++ spx-gui/src/apis/common/exception.ts
@@ -3,7 +3,7 @@
  */
 
 import type { LocaleMessage } from '@/utils/i18n'
-import { Exception } from '@/utils/error'
+import { Exception } from '@/utils/exception'
 
 export class ApiException extends Exception {
 
@@ -15,11 +15,11 @@ export class ApiException extends Exception {
     message: string
   ) {
     super(`[${code}] ${message}`)
-    this.userMessage = apiErrorMessages[this.code as ApiErrorCode] ?? null
+    this.userMessage = codeMessages[this.code as ApiExceptionCode] ?? null
   }
 }
 
-export enum ApiErrorCode {
+export enum ApiExceptionCode {
   errorInvalidArgs = 40001,
   errorUnauthorized = 40100,
   errorForbidden = 40300,
@@ -27,24 +27,24 @@ export enum ApiErrorCode {
   errorUnknown = 50000
 }
 
-const apiErrorMessages: Record<ApiErrorCode, LocaleMessage> = {
-  [ApiErrorCode.errorInvalidArgs]: {
+const codeMessages: Record<ApiExceptionCode, LocaleMessage> = {
+  [ApiExceptionCode.errorInvalidArgs]: {
     en: 'Invalid args',
     zh: '参数错误'
   },
-  [ApiErrorCode.errorUnauthorized]: {
+  [ApiExceptionCode.errorUnauthorized]: {
     en: 'You need to login first',
     zh: '请先登录'
   },
-  [ApiErrorCode.errorForbidden]: {
+  [ApiExceptionCode.errorForbidden]: {
     en: 'Permission denied',
     zh: '没有权限'
   },
-  [ApiErrorCode.errorNotFound]: {
+  [ApiExceptionCode.errorNotFound]: {
     en: 'Resource not exist',
     zh: '资源不存在'
   },
-  [ApiErrorCode.errorUnknown]: {
+  [ApiExceptionCode.errorUnknown]: {
     en: 'Something wrong with the server',
     zh: '服务器出问题了'
   }
